[fixup] Fix QtWebEngine build on Windows

Reviewed-on: https://codereview.qt-project.org/c/qt/qtwebengine-chromium/+/651993

+++ chromium/v8/src/runtime/runtime-test.cc
@@ -1197,7 +1197,7 @@ RUNTIME_FUNCTION(Runtime_SetAllocationTimeout) {
 namespace {
 
 int FixedArrayLenFromSize(int size) {
-  return std::min({(size - OFFSET_OF_DATA_START(FixedArray)) / kTaggedSize,
+  return std::min({(int)(size - OFFSET_OF_DATA_START(FixedArray)) / kTaggedSize,
                    FixedArray::kMaxRegularLength});
 }
